feat: Lotto.java 로또 당첨개수 반환 기능 추가
